# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Web Appliance Dashboard API
  description: RESTful API für die Verwaltung von Web-Appliances, Services und SSH-Verbindungen
  version: 1.0.3
  contact:
    name: Alf Lewerken
    email: support@example.com
    url: https://github.com/alflewerken/web-appliance-dashboard
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001/api
    description: Development server
  - url: https://api.appliance-dashboard.com/api
    description: Production server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login endpoint
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        username:
          type: string
          example: "admin"
        email:
          type: string
          format: email
          example: "admin@example.com"
        role:
          type: string
          enum: [admin, user]
          example: "admin"
        created_at:
          type: string
          format: date-time
        last_login:
          type: string
          format: date-time
...
  parameters:
...
    pageParam:
      name: page
      in: query
      schema:
        type: integer
        default: 1
      description: Seitennummer
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 20
      description: Einträge pro Seite
    searchParam:
      name: search
      in: query
      schema:
        type: string
      description: Suchbegriff
...
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: Alle Benutzer auflisten
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/searchParam'
      responses:
        '200':
          description: Benutzerliste
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  users:
                    type: array
                    items:
                      $ref: '#/components/schemas/User'
                  pagination:
                    type: object
                    properties:
                      total:
                        type: integer
                      page:
                        type: integer
                      pages:
                        type: integer
                      limit:
                        type: integer

    post:
      tags:
        - Users
      summary: Neuen Benutzer erstellen
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
                - email
              properties:
                username:
                  type: string
                password:
                  type: string
                  format: password
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [admin, user]
      responses:
        '201':
          description: Benutzer erstellt
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
...